Add kioclient to the Linux/BSD fallback chain after gnome-open
`kioclient exec` is KDE's equivalent to gvfs-open. In this version using
a fixed precedence order, it is added after all entries which existed in
the previous released version to ensure that it will not change the
behaviour for users of non-KDE desktops.

Were this starting a new major version, I would put kioclient ahead of
gvfs-open and gnome-open in the precedence order since it's more common to
find GTK+ applications on KDE desktops (and, thus `gvfs-open` too) than to
find KDE applications on GNOME desktops.