tests/android: Use `adb reverse` to access host webserver, allow running on !macos
Instead of chaining a bunch of Unix commands to pry out the IP address
of a random adapter, that may not be available on every distro, use the
standard `adb reverse` command to forward a specific port from the
Android phone (or emulator) to the host, where the test-webserver is
running.  This makes the test-script host-independent (should in theory
run even on Windows now) and does not require the device at test (which
could be a regular Android phone if not having an emulator running) to
be connected to / reachable over the same network.